Move public headers to snc folder
to use the project with FetchContent
