+++ commandline/genericSender2.cpp
@@ -1,4 +1,4 @@
-#include "client.h"
+#include "snc/client.h"
 #include "stdlib.h"
 #include <string>
