+++ snc/config.h
