+++ src/CMakeLists.txt
@@ -3,10 +3,6 @@ PRIVATE
   broker.cpp
   commandNames.h
   commandNames.cpp
-
-PUBLIC
-  $<BUILD_INTERFACE:${CMAKE_CURRENT_LIST_DIR}/config.h>
-  $<INSTALL_INTERFACE:include/snc/config.h>
 )
 
 target_sources(snc_client
@@ -14,8 +10,4 @@ PRIVATE
     client.cpp
     commandNames.h
     commandNames.cpp
-
-PUBLIC
-$<BUILD_INTERFACE:${CMAKE_CURRENT_LIST_DIR}/client.h>
-$<INSTALL_INTERFACE:include/snc/client.h>
 )
